+++ RELEASE-NOTES.md
@@ -2,6 +2,12 @@
 ### Version 4.3.1-EUI-3473-challenged-access-request-screen
 **EUI-3473** Challenged Access Request screen for Case Access Management
 
+### Version 4.3.1-switch-tabs
+**EUI-3881** Fix bug that it can't switch case detail tabs after inserting prepended tabs
+
+### Version 4.0.0-value-disappearing
+**EUI-4364** Values disappearing after mid event callback when modifying case data
+
 ### Version 4.1.6-EUI-4343-case-field-validation-red-bar-missing
 **EUI-4343** CCD case field validation indicator missing in WA2 release 2 part2 branch